FIX: remove init & installBusyBox when not in the docker_management directory
Former-commit-id: 64ffc94754d665ebeb5f0cece4a034bba0fb6c10

+++ docker_management/docker_image_tuxml.py
@@ -124,8 +124,9 @@ def create_sub_image_tuxml_compressed(tmp_location):
         image=NAME_BASE_IMAGE,
         path=tmp_location)
 
-    os.remove("{}/installBusyBox.sh".format(tmp_location))
-    os.remove("{}/init".format(tmp_location))
+    if tmp_location != os.path.dirname(os.path.abspath(__file__)):
+        os.remove("{}/installBusyBox.sh".format(tmp_location))
+        os.remove("{}/init".format(tmp_location))
 
 
 ## create_image_tuxml_compressed
